Parse dc_titles, try to add premis events

+++ tests/conftest.py
@@ -13,6 +13,7 @@ def pytest_addoption(parser: pytest.Parser):
 @pytest.fixture(scope="session", autouse=True)
 def clear_output(request):
     should_clear = request.config.getoption("--clear")
-    if should_clear:
+    path = Path("tests/output")
+    if should_clear and path.exists():
         shutil.rmtree("tests/output")
         Path("tests/output").mkdir()
